Split runner tests by domain
