+++ crates/webcodex-runner/src/main_tests.rs
@@ -90,552 +90,6 @@ fn runtime_config(cfg: &AgentConfig) -> Arc<ReloadableAgentConfig> {
     Arc::new(ReloadableAgentConfig::new(cfg.clone(), PathBuf::new()))
 }
 
-#[test]
-fn bounded_response_body_reader_stops_after_limit_plus_one() {
-    let mut reader = std::io::Cursor::new(vec![b'x'; 66]);
-    let body = read_bounded_response_body(&mut reader, None, 64).unwrap();
-    assert!(body.exceeded_limit);
-    assert_eq!(body.bytes.len(), 64);
-    assert_eq!(
-        reader.position(),
-        65,
-        "the bounded reader must not consume the unbounded remainder"
-    );
-}
-
-#[test]
-fn response_decode_distinguishes_empty_eof_and_complete_syntax_errors() {
-    for bytes in [b"".as_slice(), br#"{"success":true,"request":"#.as_slice()] {
-        let error = decode_json_response::<ShellAgentPollResponse>(
-            AGENT_POLL_PATH,
-            reqwest::StatusCode::OK,
-            "application/json",
-            BoundedResponseBody {
-                bytes: bytes.to_vec(),
-                exceeded_limit: false,
-            },
-        )
-        .unwrap_err();
-        assert_eq!(error.kind, AgentHttpErrorKind::DecodeTransient);
-    }
-
-    let error = decode_json_response::<ShellAgentPollResponse>(
-        AGENT_POLL_PATH,
-        reqwest::StatusCode::OK,
-        "application/json",
-        BoundedResponseBody {
-            bytes: b"{not-json".to_vec(),
-            exceeded_limit: false,
-        },
-    )
-    .unwrap_err();
-    assert_eq!(error.kind, AgentHttpErrorKind::ProtocolDecode);
-    assert!(error.summary.contains("serde_category=syntax"));
-    assert!(!error.to_string().contains("{not-json"));
-}
-
-#[test]
-fn protocol_decode_diagnostics_omit_queries_credentials_and_response_values() {
-    let content_type = reqwest::header::HeaderValue::from_static(
-        "application/json; authorization=Bearer SECRET-TOKEN",
-    );
-    let content_type = bounded_response_content_type(Some(&content_type), "SECRET-TOKEN");
-    let error = decode_json_response::<ShellAgentPollResponse>(
-        "/api/shell/agent/poll?token=SECRET-TOKEN",
-        reqwest::StatusCode::OK,
-        &content_type,
-        BoundedResponseBody {
-            bytes: br#"{"success":"SECRET-TOKEN","request":null,"error":null}"#.to_vec(),
-            exceeded_limit: false,
-        },
-    )
-    .unwrap_err();
-    let message = error.to_string();
-    assert_eq!(error.kind, AgentHttpErrorKind::ProtocolDecode);
-    assert!(
-        message.contains("content_type=application/json"),
-        "{message}"
-    );
-    assert!(!message.contains('?'), "{message}");
-    assert!(!message.contains("SECRET-TOKEN"), "{message}");
-    assert!(!message.contains("authorization"), "{message}");
-    assert!(!message.contains('\n'), "{message}");
-}
-
-#[test]
-fn result_400_is_classified_permanent_with_bounded_structured_reason() {
-    let error = AgentHttpError::status(
-        "/api/shell/agent/result",
-        reqwest::StatusCode::BAD_REQUEST,
-        r#"{"success":false,"error":"unknown or expired shell request: req-1"}"#,
-    );
-    assert_eq!(error.kind, AgentHttpErrorKind::ClientRejected);
-    let message = error.to_string();
-    assert!(
-        message.contains("server rejected /api/shell/agent/result request"),
-        "{message}"
-    );
-    assert!(message.contains("HTTP 400 Bad Request"), "{message}");
-    assert!(
-        message.contains("unknown or expired shell request: req-1"),
-        "{message}"
-    );
-}
-
-#[test]
-fn result_4xx_html_bodies_stay_permanent_and_never_leak_markup() {
-    let bad_request = AgentHttpError::status(
-        "/api/shell/agent/result",
-        reqwest::StatusCode::BAD_REQUEST,
-        "<html>\n<body><h1>400 Bad Request</h1><center>nginx</center></body>\n</html>",
-    );
-    assert_eq!(bad_request.kind, AgentHttpErrorKind::ClientRejected);
-    assert!(!bad_request.to_string().contains("<html"), "{bad_request}");
-
-    let too_large = AgentHttpError::status(
-        "/api/shell/agent/result",
-        reqwest::StatusCode::PAYLOAD_TOO_LARGE,
-        "<html><center>nginx</center><center>413 Request Entity Too Large</center></html>",
-    );
-    assert_eq!(too_large.kind, AgentHttpErrorKind::ClientRejected);
-    assert!(!too_large.to_string().contains("nginx"), "{too_large}");
-}
-
-#[test]
-fn result_400_structured_reason_is_bounded_for_large_json_bodies() {
-    let huge = format!(r#"{{"success":false,"error":"{}"}}"#, "x".repeat(10_000));
-    let error = AgentHttpError::status(
-        "/api/shell/agent/result",
-        reqwest::StatusCode::BAD_REQUEST,
-        &huge,
-    );
-    assert_eq!(error.kind, AgentHttpErrorKind::ClientRejected);
-    let message = error.to_string();
-    assert!(
-        message.chars().count() < 300,
-        "unbounded message: {} chars",
-        message.chars().count()
-    );
-}
-
-#[test]
-fn http_status_classification_keeps_retryable_auth_and_gateway_kinds() {
-    let cases = [
-        (reqwest::StatusCode::UNAUTHORIZED, AgentHttpErrorKind::Auth),
-        (reqwest::StatusCode::FORBIDDEN, AgentHttpErrorKind::Auth),
-        (reqwest::StatusCode::NOT_FOUND, AgentHttpErrorKind::NotFound),
-        (
-            reqwest::StatusCode::REQUEST_TIMEOUT,
-            AgentHttpErrorKind::Status,
-        ),
-        (
-            reqwest::StatusCode::TOO_MANY_REQUESTS,
-            AgentHttpErrorKind::Status,
-        ),
-        (
-            reqwest::StatusCode::INTERNAL_SERVER_ERROR,
-            AgentHttpErrorKind::ServerUnavailable,
-        ),
-        (
-            reqwest::StatusCode::BAD_GATEWAY,
-            AgentHttpErrorKind::ServerUnavailable,
-        ),
-        (
-            reqwest::StatusCode::SERVICE_UNAVAILABLE,
-            AgentHttpErrorKind::ServerUnavailable,
-        ),
-        (
-            reqwest::StatusCode::GATEWAY_TIMEOUT,
-            AgentHttpErrorKind::ServerUnavailable,
-        ),
-    ];
-    for (status, expected) in cases {
-        let error = AgentHttpError::status("/api/shell/agent/result", status, "{}");
-        assert_eq!(error.kind, expected, "status {status}");
-    }
-}
-
-#[test]
-fn register_recovery_classification_is_strict_about_lease_conflicts() {
-    let lease = AgentHttpError::status(
-        AGENT_REGISTER_PATH,
-        reqwest::StatusCode::BAD_REQUEST,
-        r#"{"success":false,"error":"agent client oe is already online with a different instance"}"#,
-    );
-    let lease = RegisterError::from_http(lease, "oe");
-    assert_eq!(
-        lease.recovery_action(),
-        RegisterRecoveryAction::WaitForLease
-    );
-
-    for body in [
-        r#"{"success":false,"error":"agent client identity is unavailable"}"#,
-        r#"{"success":false,"error":"agent token owner is 'alice'; cannot register owner 'bob'"}"#,
-        r#"{"success":false,"error":"agent client oe is already online"}"#,
-    ] {
-        let rejected =
-            AgentHttpError::status(AGENT_REGISTER_PATH, reqwest::StatusCode::BAD_REQUEST, body);
-        let rejected = RegisterError::from_http(rejected, "oe");
-        assert_eq!(
-            rejected.recovery_action(),
-            RegisterRecoveryAction::Fatal,
-            "{body}"
-        );
-    }
-}
-
-#[test]
-fn poll_recovery_actions_separate_transport_session_and_fatal_errors() {
-    let transient = PollError::from_http(
-        AgentHttpError::status(
-            AGENT_POLL_PATH,
-            reqwest::StatusCode::INTERNAL_SERVER_ERROR,
-            "{}",
-        ),
-        "oe",
-    );
-    assert_eq!(
-        transient.recovery_action(),
-        PollingRecoveryAction::RetryPoll
-    );
-
-    let missing_session = PollError::from_http(
-        AgentHttpError::status(
-            AGENT_POLL_PATH,
-            reqwest::StatusCode::BAD_REQUEST,
-            r#"{"success":false,"error":"unknown shell client: oe"}"#,
-        ),
-        "oe",
-    );
-    assert_eq!(
-        missing_session.recovery_action(),
-        PollingRecoveryAction::ReRegister
-    );
-
-    let ordinary_400 = PollError::from_http(
-        AgentHttpError::status(
-            AGENT_POLL_PATH,
-            reqwest::StatusCode::BAD_REQUEST,
-            r#"{"success":false,"error":"invalid poll payload"}"#,
-        ),
-        "oe",
-    );
-    assert_eq!(ordinary_400.recovery_action(), PollingRecoveryAction::Fatal);
-}
-
-#[test]
-fn tls_configuration_markers_are_fatal_but_dns_and_eof_are_not() {
-    assert!(looks_like_fatal_tls_request(
-        "error: invalid peer certificate: UnknownIssuer"
-    ));
-    assert!(looks_like_fatal_tls_request(
-        "tls error: no application protocol; ALPN mismatch"
-    ));
-    assert!(!looks_like_fatal_tls_request(
-        "dns error: temporary failure in name resolution"
-    ));
-    assert!(!looks_like_fatal_tls_request(
-        "connection closed: unexpected EOF"
-    ));
-}
-
-#[test]
-fn submit_fatal_error_classes_map_to_terminal_poll_contract() {
-    assert!(PollError::from_submit(SubmitResultError::FatalAuth("auth".into())).is_terminal());
-    assert!(
-        PollError::from_submit(SubmitResultError::FatalProtocol("missing".into())).is_terminal()
-    );
-    assert!(PollError::from_submit(SubmitResultError::FatalConfig("tls".into())).is_terminal());
-    assert!(
-        PollError::from_submit(SubmitResultError::TransportClosed("closed".into())).is_terminal()
-    );
-    let shutdown = PollError::from_submit(SubmitResultError::Shutdown("process shutdown".into()));
-    assert!(!shutdown.is_terminal());
-    assert!(shutdown.is_shutdown());
-}
-
-fn reload_toml(
-    client_id: &str,
-    max_jobs: Option<usize>,
-    max_timeout: u64,
-    max_output: usize,
-    shell_program: &str,
-    strategy: &str,
-    claude_enabled: bool,
-    claude_command: &str,
-    search_mapping: &str,
-) -> String {
-    let max_jobs = max_jobs
-        .map(|value| format!("max_concurrent_jobs = {value}\n"))
-        .unwrap_or_default();
-    format!(
-        r#"server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "{client_id}"
-owner = "alice"
-poll_interval_ms = 1000
-{max_jobs}
-# Explicit projects_dir: load_config materializes the default from the
-# per-user config base, which depends on ambient HOME/USERPROFILE that other
-# tests mutate.
-projects_dir = "projects.d"
-policy.allow_raw_shell = true
-policy.allow_cwd_anywhere = true
-policy.allowed_roots = ["/"]
-policy.max_timeout_secs = {max_timeout}
-policy.max_output_bytes = {max_output}
-shell.program = "{shell_program}"
-shell.args = ["-c"]
-tool_providers.strategy = "{strategy}"
-tool_providers.claude_code.enabled = {claude_enabled}
-tool_providers.claude_code.command = "{claude_command}"
-tool_providers.claude_code.args = ["mcp", "serve"]
-tool_providers.claude_code.timeout_secs = 30
-[tool_providers.claude_code.mapping]
-search_project_text = "{search_mapping}"
-"#
-    )
-}
-
-fn reload_fixture() -> (tempfile::TempDir, PathBuf, ReloadableAgentConfig) {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        reload_toml(
-            "oe",
-            None,
-            60,
-            1024,
-            "sh",
-            "native",
-            false,
-            "claude",
-            "project_search_generation_1",
-        ),
-    )
-    .unwrap();
-    let runtime = ReloadableAgentConfig::new(load_config(&path).unwrap(), path.clone());
-    (tmp, path, runtime)
-}
-
-#[test]
-fn reload_field_classification_is_exhaustive_and_allowlisted() {
-    let startup = test_config(PathBuf::from("projects-a"));
-    let mut hot_only = startup.clone();
-    hot_only.policy.max_timeout_secs += 1;
-    hot_only.shell.program = "bash".to_string();
-    hot_only.tool_providers.strategy =
-        webcodex_runner::config::ToolProviderStrategy::ClaudeCodeThenNative;
-    assert!(webcodex_runner::config::restart_required_fields(&startup, &hot_only).is_empty());
-
-    let mut changed = hot_only;
-    changed.server_url.push_str("/other");
-    changed.token.push('2');
-    changed.client_id.push('2');
-    changed.display_name = Some("changed".to_string());
-    changed.owner = Some("changed".to_string());
-    changed.hostname = Some("changed".to_string());
-    changed.host_context = Some(shell_protocol::AgentHostContext {
-        role: Some("primary_development".to_string()),
-        runtime: Some("Prefer this Runner for ordinary development.".to_string()),
-        ..Default::default()
-    });
-    changed.projects_dir = Some(PathBuf::from("projects-b"));
-    changed.temporary_projects_root = Some(PathBuf::from("/tmp/webcodex-temporary"));
-    changed.poll_interval_ms += 1;
-    changed.capabilities = Some(ShellClientCapabilities::default());
-    changed.max_concurrent_jobs = Some(4);
-    changed.transport = Some(TRANSPORT_QUIC.to_string());
-    changed.websocket_connect_timeout_secs += 1;
-    changed.quic = Some(quic_client_config());
-    assert_eq!(
-            webcodex_runner::config::restart_required_fields(&startup, &changed).join(" "),
-            "capabilities client_id display_name hostname host_context max_concurrent_jobs owner poll_interval_ms projects_dir temporary_projects_root quic server_url token transport websocket_connect_timeout_secs"
-        );
-}
-
-#[test]
-fn valid_reload_switches_one_complete_generation_and_preserves_old_snapshot() {
-    let (_tmp, path, runtime) = reload_fixture();
-    let old = runtime.snapshot();
-
-    assert_eq!(
-        old.external_tools.configured_search_tool_name(),
-        Some("project_search_generation_1")
-    );
-    assert_eq!(
-        old.external_tools.status().claude_code.process_state,
-        "not_started"
-    );
-
-    std::fs::write(
-        &path,
-        reload_toml(
-            "oe",
-            None,
-            120,
-            2048,
-            "bash",
-            "claude_code_then_native",
-            false,
-            "claude",
-            "project_search_generation_2",
-        ),
-    )
-    .unwrap();
-    let status = runtime.reload();
-    let new = runtime.snapshot();
-
-    assert_eq!(status.last_reload_result, "success");
-    assert_eq!(status.generation, 2);
-    assert!(!status.restart_required);
-    assert_eq!(
-        (
-            old.generation,
-            old.policy.max_timeout_secs,
-            old.shell.program.as_str()
-        ),
-        (1, 60, "sh")
-    );
-    assert_eq!(old.external_tools.status().strategy, "native");
-    assert_eq!(
-        old.external_tools.configured_search_tool_name(),
-        Some("project_search_generation_1")
-    );
-    assert_eq!(
-        (
-            new.policy.max_timeout_secs,
-            new.policy.max_output_bytes,
-            new.shell.program.as_str()
-        ),
-        (120, 2048, "bash")
-    );
-    assert_eq!(
-        new.external_tools.status().strategy,
-        "claude_code_then_native"
-    );
-    assert_eq!(
-        new.external_tools.configured_search_tool_name(),
-        Some("project_search_generation_2")
-    );
-    assert_eq!(
-        new.external_tools.status().claude_code.process_state,
-        "not_started"
-    );
-}
-
-#[test]
-fn failed_reload_keeps_generation_and_can_recover() {
-    let (_tmp, path, runtime) = reload_fixture();
-    let old = runtime.snapshot();
-
-    std::fs::remove_file(&path).unwrap();
-    let status = runtime.reload();
-    assert_eq!(status.generation, 1);
-    assert_eq!(
-        status.last_reload_error_code.as_deref(),
-        Some("config_read_failed")
-    );
-
-    for (candidate, code) in [
-        ("{ invalid toml".to_string(), "config_parse_failed"),
-        (
-            reload_toml(
-                "oe",
-                None,
-                60,
-                1024,
-                "",
-                "native",
-                false,
-                "claude",
-                "project_search_generation_1",
-            ),
-            "config_validation_failed",
-        ),
-        (
-            reload_toml(
-                "oe",
-                None,
-                60,
-                1024,
-                "sh",
-                "native",
-                true,
-                "",
-                "project_search_generation_1",
-            ),
-            "provider_config_invalid",
-        ),
-    ] {
-        std::fs::write(&path, candidate).unwrap();
-        let status = runtime.reload();
-        assert_eq!(status.generation, 1);
-        assert_eq!(status.last_reload_result, "failure");
-        assert_eq!(status.last_reload_error_code.as_deref(), Some(code));
-    }
-    assert_eq!(old.policy.max_timeout_secs, 60);
-    let serialized = serde_json::to_string(&runtime.snapshot().reload_status()).unwrap();
-    assert!(!serialized.contains(path.to_string_lossy().as_ref()));
-    assert!(!serialized.contains("test-token"));
-
-    std::fs::write(
-        &path,
-        reload_toml(
-            "oe",
-            None,
-            90,
-            1024,
-            "sh",
-            "native",
-            false,
-            "claude",
-            "project_search_generation_2",
-        ),
-    )
-    .unwrap();
-    assert_eq!(runtime.reload().generation, 2);
-    assert_eq!(runtime.snapshot().policy.max_timeout_secs, 90);
-}
-
-#[test]
-fn mixed_reload_applies_hot_fields_and_reports_static_restart_fields() {
-    let (_tmp, path, runtime) = reload_fixture();
-    std::fs::write(
-        &path,
-        reload_toml(
-            "oe-new",
-            Some(8),
-            180,
-            4096,
-            "bash",
-            "native",
-            false,
-            "claude",
-            "project_search_generation_2",
-        ),
-    )
-    .unwrap();
-
-    let status = runtime.reload();
-    let active = runtime.snapshot();
-    assert_eq!(status.last_reload_result, "partial");
-    assert!(status.restart_required);
-    assert_eq!(
-        status.restart_required_fields,
-        ["client_id", "max_concurrent_jobs"]
-    );
-    assert_eq!(
-        (
-            active.policy.max_timeout_secs,
-            active.policy.max_output_bytes,
-            active.shell.program.as_str()
-        ),
-        (180, 4096, "bash")
-    );
-}
-
 fn quic_client_config() -> QuicClientConfig {
     QuicClientConfig {
         server_addr: "v4.example.test:8443".to_string(),
@@ -646,766 +100,12 @@ fn quic_client_config() -> QuicClientConfig {
     }
 }
 
-#[test]
-fn agent_config_defaults_transport_to_websocket_without_quic_section() {
-    // No transport field and no [quic] section: default stays websocket.
-    let toml = r#"
-server_url = "http://127.0.0.1:8000"
-token = "t"
-client_id = "oe"
-"#;
-    let cfg: AgentConfig = toml::from_str(toml).unwrap();
-    assert!(cfg.transport.is_none());
-    assert!(cfg.quic.is_none());
-    assert_eq!(effective_transport(&cfg), TRANSPORT_WEBSOCKET);
-    assert_eq!(
-        cfg.websocket_connect_timeout_secs,
-        default_websocket_connect_timeout_secs()
-    );
-    assert_eq!(
-        auto_transport_plan(&cfg),
-        vec![TRANSPORT_WEBSOCKET, TRANSPORT_POLLING]
-    );
-}
-
-#[test]
-fn agent_config_rejects_zero_websocket_connect_timeout() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "t"
-client_id = "oe"
-websocket_connect_timeout_secs = 0
-"#,
-    )
-    .unwrap();
-
-    let err = load_config(&path).unwrap_err();
-    assert!(
-        err.contains("websocket_connect_timeout_secs must be > 0"),
-        "{err}"
-    );
-}
-
-#[test]
-fn agent_config_rejects_relative_temporary_projects_root() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "t"
-client_id = "oe"
-temporary_projects_root = "temporary"
-"#,
-    )
-    .unwrap();
-
-    let err = load_config(&path).unwrap_err();
-    assert!(
-        err.contains("temporary_projects_root must be a non-empty absolute path"),
-        "{err}"
-    );
-}
-
-#[test]
-fn agent_config_accepts_transport_quic_with_quic_section() {
-    let toml = r#"
-server_url = "http://127.0.0.1:8000"
-token = "t"
-client_id = "oe"
-transport = "quic"
-
-[quic]
-server_addr = "v4.example.test:8443"
-server_name = "v4.example.test"
-"#;
-    let cfg: AgentConfig = toml::from_str(toml).unwrap();
-    assert_eq!(cfg.transport.as_deref(), Some("quic"));
-    let quic = cfg.quic.expect("quic section");
-    assert_eq!(quic.server_addr, "v4.example.test:8443");
-    assert_eq!(quic.server_name, "v4.example.test");
-    // Defaults applied.
-    assert_eq!(quic.alpn, "webcodex-runner/1");
-    assert_eq!(quic.connect_timeout_secs, 10);
-    assert_eq!(quic.keepalive_interval_secs, 20);
-}
-
-#[test]
-fn agent_config_accepts_transport_auto() {
-    let toml = r#"
-server_url = "http://127.0.0.1:8000"
-token = "t"
-client_id = "oe"
-transport = "auto"
-"#;
-    let cfg: AgentConfig = toml::from_str(toml).unwrap();
-    assert_eq!(cfg.transport.as_deref(), Some(TRANSPORT_AUTO));
-    assert_eq!(effective_transport(&cfg), TRANSPORT_AUTO);
-    assert_eq!(
-        auto_transport_plan(&cfg),
-        vec![TRANSPORT_WEBSOCKET, TRANSPORT_POLLING]
-    );
-}
-
-#[test]
-fn auto_transport_plan_tries_quic_then_websocket_then_polling() {
-    let mut cfg = test_config(PathBuf::from("/tmp/x"));
-    cfg.transport = Some(TRANSPORT_AUTO.to_string());
-    cfg.quic = Some(quic_client_config());
-    assert_eq!(
-        auto_transport_plan(&cfg),
-        vec![TRANSPORT_QUIC, TRANSPORT_WEBSOCKET, TRANSPORT_POLLING]
-    );
-}
-
-#[test]
-fn strict_quic_transport_still_requires_quic_section() {
-    let mut cfg = test_config(PathBuf::from("/tmp/x"));
-    cfg.transport = Some(TRANSPORT_QUIC.to_string());
-    let err = resolve_quic_config(&cfg).unwrap_err();
-    assert!(err.contains("transport=quic requires a [quic] section"));
-    assert_eq!(effective_transport(&cfg), TRANSPORT_QUIC);
-}
-
-#[test]
-fn resolve_quic_config_errors_when_section_missing() {
-    let mut cfg = test_config(PathBuf::from("/tmp/x"));
-    cfg.transport = Some("quic".to_string());
-    let err = resolve_quic_config(&cfg).unwrap_err();
-    assert!(err.contains("[quic]"), "err was: {err}");
-}
-
-#[test]
-fn resolve_quic_config_errors_when_server_addr_or_name_missing() {
-    let mut cfg = test_config(PathBuf::from("/tmp/x"));
-    cfg.transport = Some("quic".to_string());
-
-    // Missing server_addr.
-    cfg.quic = Some(QuicClientConfig {
-        server_addr: "  ".to_string(),
-        server_name: "v4.example.test".to_string(),
-        alpn: default_quic_alpn(),
-        connect_timeout_secs: 10,
-        keepalive_interval_secs: 20,
-    });
-    let err = resolve_quic_config(&cfg).unwrap_err();
-    assert!(err.contains("server_addr"), "err was: {err}");
-
-    // Missing server_name.
-    cfg.quic = Some(QuicClientConfig {
-        server_addr: "v4.example.test:8443".to_string(),
-        server_name: String::new(),
-        alpn: default_quic_alpn(),
-        connect_timeout_secs: 10,
-        keepalive_interval_secs: 20,
-    });
-    let err = resolve_quic_config(&cfg).unwrap_err();
-    assert!(err.contains("server_name"), "err was: {err}");
-}
-
-#[test]
-fn resolve_quic_config_accepts_valid_section() {
-    let mut cfg = test_config(PathBuf::from("/tmp/x"));
-    cfg.transport = Some("quic".to_string());
-    cfg.quic = Some(quic_client_config());
-    let resolved = resolve_quic_config(&cfg).unwrap();
-    assert_eq!(resolved.server_addr, "v4.example.test:8443");
-    assert_eq!(resolved.server_name, "v4.example.test");
-}
-
-#[test]
-fn resolve_quic_server_addrs_accepts_hostname_port() {
-    let addrs = resolve_quic_server_addrs("localhost:8443").unwrap();
-    assert!(addrs.iter().any(|addr| addr.port() == 8443));
-}
-
-#[test]
-fn resolve_quic_server_addrs_rejects_missing_port() {
-    let err = resolve_quic_server_addrs("localhost").unwrap_err();
-    assert!(err.contains("failed to resolve"), "err was: {err}");
-}
-
-#[test]
-fn quic_client_bind_addr_matches_remote_address_family() {
-    let v4: SocketAddr = "127.0.0.1:8443".parse().unwrap();
-    let v6: SocketAddr = "[::1]:8443".parse().unwrap();
-    assert!(quic_client_bind_addr_for(v4).is_ipv4());
-    assert!(quic_client_bind_addr_for(v6).is_ipv6());
-}
-
-#[test]
-fn agent_cli_help_and_version_exit_before_runtime() {
-    match parse_agent_args(["--help"]).unwrap() {
-        AgentCliAction::Exit {
-            code,
-            stdout,
-            stderr,
-        } => {
-            assert_eq!(code, 0);
-            assert!(stdout.contains("Usage: webcodex-runner"));
-            assert!(!stdout.contains("webcodex-runner init"));
-            assert!(stderr.is_empty());
-        }
-        other => panic!("expected help exit, got {other:?}"),
-    }
-    match parse_agent_args(["--version"]).unwrap() {
-        AgentCliAction::Exit {
-            code,
-            stdout,
-            stderr,
-        } => {
-            assert_eq!(code, 0);
-            assert!(stdout.starts_with(&format!(
-                "webcodex-runner {} (commit ",
-                env!("CARGO_PKG_VERSION")
-            )));
-            assert!(stdout.trim_end().ends_with(')'));
-            assert_ne!(
-                stdout,
-                format!("webcodex-runner {}\n", env!("CARGO_PKG_VERSION"))
-            );
-            assert!(stderr.is_empty());
-        }
-        other => panic!("expected version exit, got {other:?}"),
-    }
-}
-
-#[test]
-fn agent_cli_has_no_init_alias() {
-    let error = parse_agent_args(["init"]).unwrap_err();
-    assert!(error.contains("unknown argument: init"));
-}
-
-#[test]
-fn agent_version_output_includes_build_metadata() {
-    match parse_agent_args(["-V"]).unwrap() {
-        AgentCliAction::Exit {
-            code,
-            stdout,
-            stderr,
-        } => {
-            assert_eq!(code, 0);
-            assert!(stdout.contains("commit "));
-            assert!(stdout.starts_with("webcodex-runner "));
-            assert!(stderr.is_empty());
-        }
-        other => panic!("expected version exit, got {other:?}"),
-    }
-}
-
-#[test]
-fn agent_cli_legacy_runtime_args_are_preserved() {
-    let action = parse_agent_args(["--config", "/tmp/agent.toml", "--once"]).unwrap();
-    assert_eq!(
-        action,
-        AgentCliAction::Run {
-            config_path: PathBuf::from("/tmp/agent.toml"),
-            once: true,
-        }
-    );
-}
-
-#[test]
-fn agent_cli_profile_derives_default_config_path() {
-    let action = parse_agent_args(["--profile", "special"]).unwrap();
-    assert_eq!(
-        action,
-        AgentCliAction::Run {
-            config_path: client_profile_agent_config("special").unwrap(),
-            once: false,
-        }
-    );
-}
-
-#[test]
-fn agent_cli_explicit_config_overrides_profile() {
-    let action = parse_agent_args(["--profile", "special", "--config", "/tmp/agent.toml"]).unwrap();
-    assert_eq!(
-        action,
-        AgentCliAction::Run {
-            config_path: PathBuf::from("/tmp/agent.toml"),
-            once: false,
-        }
-    );
-}
-
-#[test]
-fn agent_cli_rejects_unsafe_profile() {
-    let err = parse_agent_args(["--profile", "../x"]).unwrap_err();
-    assert_eq!(err, CLIENT_PROFILE_ERROR);
-}
-
-#[test]
-fn empty_tokens_config_parser_accepts_empty_and_whitespace_token() {
-    for token in ["", "   "] {
-        let tmp = tempfile::tempdir().unwrap();
-        let path = tmp.path().join("agent.toml");
-        std::fs::write(
-                &path,
-                format!(
-                    "server_url = \"http://127.0.0.1:8000\"\ntoken = \"{}\"\nclient_id = \"open-agent\"\n[policy]\nallow_cwd_anywhere = true\n",
-                    token
-                ),
-            )
-            .unwrap();
-
-        let cfg = load_config(&path).unwrap();
-        assert_eq!(cfg.token, token);
-        assert_eq!(non_empty_token(&cfg.token), None);
-    }
-}
-
-#[test]
-fn agent_config_host_context_is_normalized_closed_and_restart_scoped() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-projects_dir = "projects.d"
-
-[host_context]
-role = " server_host "
-runtime = " Prefer this Runner for Server-host operations. "
-service = "Use the ordinary host-local service mechanism."
-
-[policy]
-allow_cwd_anywhere = true
-"#,
-    )
-    .unwrap();
-    let cfg = load_config(&path).unwrap();
-    let context = cfg.host_context.as_ref().expect("host context");
-    assert_eq!(context.role.as_deref(), Some("server_host"));
-    assert_eq!(
-        context.runtime.as_deref(),
-        Some("Prefer this Runner for Server-host operations.")
-    );
-
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-projects_dir = "projects.d"
-[host_context]
-role = "server_host"
-arbitrary = "not allowed"
-[policy]
-allow_cwd_anywhere = true
-"#,
-    )
-    .unwrap();
-    let err = load_config(&path).unwrap_err();
-    assert!(err.contains("failed to parse config"), "{err}");
-    assert!(err.contains("arbitrary"), "{err}");
-}
-
-#[test]
-fn agent_config_without_shell_section_parses() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[policy]
-allow_cwd_anywhere = true
-"#,
-    )
-    .unwrap();
-
-    let cfg = load_config(&path).unwrap();
-    assert_eq!(cfg.shell, ShellConfig::default());
-    assert_eq!(cfg.shell.max_persistent_shells, 8);
-    assert_eq!(cfg.shell.persistent_shell_idle_timeout_secs, 30 * 60);
-}
-
-#[test]
-fn agent_config_persistent_shell_limits_are_validated() {
-    let mut shell = ShellConfig {
-        max_persistent_shells: 0,
-        ..Default::default()
-    };
-    assert!(validate_shell_config(&shell)
-        .unwrap_err()
-        .contains("max_persistent_shells"));
-
-    shell.max_persistent_shells = 65;
-    assert!(validate_shell_config(&shell)
-        .unwrap_err()
-        .contains("max_persistent_shells"));
-
-    shell.max_persistent_shells = 8;
-    shell.persistent_shell_idle_timeout_secs = 0;
-    assert!(validate_shell_config(&shell)
-        .unwrap_err()
-        .contains("persistent_shell_idle_timeout_secs"));
-
-    shell.persistent_shell_idle_timeout_secs = 86_401;
-    assert!(validate_shell_config(&shell)
-        .unwrap_err()
-        .contains("persistent_shell_idle_timeout_secs"));
-}
-
-#[test]
-fn agent_config_loads_named_ssh_resources_without_authentication_material() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[ssh.resources.tmp]
-host = "tmp"
-default_cwd = "/opt/webcodex-edge"
-
-[ssh.resources.no_default]
-host = "ops-alias"
-"#,
-    )
-    .unwrap();
-
-    let cfg = load_config(&path).unwrap();
-    let tmp = cfg.ssh.resources.get("tmp").unwrap();
-    assert_eq!(tmp.host, "tmp");
-    assert_eq!(tmp.default_cwd.as_deref(), Some("/opt/webcodex-edge"));
-    assert_eq!(
-        cfg.ssh
-            .resources
-            .get("no_default")
-            .and_then(|resource| resource.default_cwd.as_deref()),
-        None
-    );
-}
-
-#[test]
-fn agent_config_shell_profiles_parse() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[policy]
-allow_cwd_anywhere = true
-
-[shell]
-default_profile = "rust"
-
-[shell.profiles.rust]
-description = "Rust development tools"
-program = "sh"
-args = ["-c"]
-init_script = '''
-export RUST_BACKTRACE=1
-'''
-
-[shell.profiles.rust.env]
-PATH = "/root/.cargo/bin:/usr/bin:/bin"
-CARGO_HOME = "/root/.cargo"
-RUSTUP_HOME = "/root/.rustup"
-
-[shell.profiles.py-venv]
-description = "Project-local Python virtual environment"
-program = "bash"
-args = ["-lc"]
-init_script = '''
-source .venv/bin/activate
-'''
-"#,
-    )
-    .unwrap();
-
-    let cfg = load_config(&path).unwrap();
-    assert_eq!(cfg.shell.default_profile.as_deref(), Some("rust"));
-    let rust = cfg.shell.profiles.get("rust").unwrap();
-    assert_eq!(rust.description.as_deref(), Some("Rust development tools"));
-    assert_eq!(rust.program.as_deref(), Some("sh"));
-    assert_eq!(rust.args.as_ref().unwrap(), &vec!["-c".to_string()]);
-    assert_eq!(
-        rust.env.get("CARGO_HOME").map(String::as_str),
-        Some("/root/.cargo")
-    );
-    assert!(rust
-        .init_script
-        .as_deref()
-        .unwrap()
-        .contains("RUST_BACKTRACE=1"));
-    assert!(cfg.shell.profiles.contains_key("py-venv"));
-}
-
-#[test]
-fn agent_config_shell_default_profile_must_exist() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[policy]
-allow_cwd_anywhere = true
-
-[shell]
-default_profile = "missing"
-
-[shell.profiles.rust]
-program = "sh"
-"#,
-    )
-    .unwrap();
-
-    let err = load_config(&path).unwrap_err();
-    assert!(err.contains("default_profile"), "{err}");
-    assert!(err.contains("missing"), "{err}");
-}
-
-#[test]
-fn agent_config_shell_profile_name_must_be_safe() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[policy]
-allow_cwd_anywhere = true
-
-[shell.profiles."bad/name"]
-program = "sh"
-"#,
-    )
-    .unwrap();
-
-    let err = load_config(&path).unwrap_err();
-    assert!(err.contains("shell profile name"), "{err}");
-    assert!(err.contains("slash"), "{err}");
-}
-
-#[test]
-fn agent_config_shell_profile_type_errors_are_reported() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[policy]
-allow_cwd_anywhere = true
-
-[shell.profiles.rust]
-args = "-c"
-"#,
-    )
-    .unwrap();
-
-    let err = load_config(&path).unwrap_err();
-    assert!(err.contains("failed to parse config"), "{err}");
-    assert!(err.contains("args"), "{err}");
-}
-
-#[test]
-fn agent_config_shell_profile_env_type_errors_are_reported() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    std::fs::write(
-        &path,
-        r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[policy]
-allow_cwd_anywhere = true
-
-[shell.profiles.rust.env]
-PATH = ["/root/.cargo/bin"]
-"#,
-    )
-    .unwrap();
-
-    let err = load_config(&path).unwrap_err();
-    assert!(err.contains("failed to parse config"), "{err}");
-    assert!(err.contains("env"), "{err}");
-}
-
-#[test]
-fn agent_config_shell_errors_do_not_include_init_script_body() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("agent.toml");
-    let secret = "DO_NOT_LEAK_THIS_INLINE_SCRIPT_BODY";
-    std::fs::write(
-        &path,
-        format!(
-            r#"
-server_url = "http://127.0.0.1:8000"
-token = "test-token"
-client_id = "agent-1"
-
-[policy]
-allow_cwd_anywhere = true
-
-[shell]
-default_profile = "missing"
-
-[shell.profiles.rust]
-init_script = '''
-export SECRET={}
-'''
-"#,
-            secret
-        ),
-    )
-    .unwrap();
-
-    let err = load_config(&path).unwrap_err();
-    assert!(err.contains("default_profile"), "{err}");
-    assert!(!err.contains(secret), "{err}");
-}
-
-#[test]
-fn agent_project_toml_parse_sorts_hook_names() {
-    let project = parse_agent_project_toml(
-        r#"
-id = "webcodex"
-path = "/root/git/webcodex"
-kind = "rust"
-shell_profile = "rust"
-
-[hooks]
-precommit = ["cargo test"]
-doctor = ["git status --short"]
-"#,
-    )
-    .unwrap();
-    let summary = agent_project_summary(&project, 123456, false);
-    assert_eq!(summary.id, "webcodex");
-    assert_eq!(summary.name.as_deref(), Some("webcodex"));
-    assert_eq!(summary.path, "/root/git/webcodex");
-    assert_eq!(summary.kind.as_deref(), Some("rust"));
-    assert_eq!(summary.hooks, vec!["doctor", "precommit"]);
-    assert_eq!(summary.updated_at, 123456);
-    assert_eq!(summary.git_branch, None);
-    assert_eq!(project.shell_profile.as_deref(), Some("rust"));
-}
-
-#[test]
-fn agent_project_toml_rejects_invalid_id() {
-    let err = parse_agent_project_toml(
-        r#"
-id = "bad id"
-path = "/tmp/webcodex"
-"#,
-    )
-    .unwrap_err();
-    assert!(err.contains("ASCII letters"));
-}
-
-#[test]
-fn agent_project_toml_hints_when_server_projects_format_is_used() {
-    let err = parse_agent_project_toml(
-        r#"
-[projects.smoke]
-path = "/root/webcodex-smoke"
-"#,
-    )
-    .unwrap_err();
-    assert!(err.contains("missing field"), "{err}");
-    assert!(err.contains("server projects.toml"), "{err}");
-    assert!(
-        err.contains("Agent projects.d files must use top-level fields"),
-        "{err}"
-    );
-    assert!(err.contains("id = \"smoke\""), "{err}");
-    assert!(err.contains("path = \"/path/to/repo\""), "{err}");
-}
-
-#[test]
-fn agent_project_toml_rejects_invalid_shell_profile() {
-    let err = parse_agent_project_toml(
-        r#"
-id = "demo"
-path = "/tmp/webcodex"
-shell_profile = "../rust"
-"#,
-    )
-    .unwrap_err();
-    assert!(err.contains("project.shell_profile"), "{err}");
-}
-
-#[test]
-fn missing_projects_dir_returns_empty_list() {
-    let tmp = tempfile::tempdir().unwrap();
-    let missing = tmp.path().join("missing-projects.d");
-    let projects = load_agent_project_summaries_from_dir(&missing);
-    assert!(projects.is_empty());
-}
-
-#[test]
-fn phase_e2_max_concurrent_jobs_normalizes_to_inventory_capacity() {
-    let tmp = tempfile::tempdir().unwrap();
-    let mut cfg = test_config(tmp.path().join("config/projects.d"));
-    assert_eq!(DEFAULT_MAX_CONCURRENT_JOBS, 4);
-    assert_eq!(max_concurrent_jobs(&cfg), DEFAULT_MAX_CONCURRENT_JOBS);
-
-    cfg.max_concurrent_jobs = Some(0);
-    assert_eq!(max_concurrent_jobs(&cfg), 1);
-
-    cfg.max_concurrent_jobs = Some(1);
-    assert_eq!(max_concurrent_jobs(&cfg), 1);
-
-    cfg.max_concurrent_jobs = Some(4);
-    assert_eq!(max_concurrent_jobs(&cfg), 4);
-
-    cfg.max_concurrent_jobs = Some(8);
-    assert_eq!(max_concurrent_jobs(&cfg), 8);
-
-    cfg.max_concurrent_jobs = Some(64);
-    assert_eq!(max_concurrent_jobs(&cfg), 64);
-
-    cfg.max_concurrent_jobs = Some(65);
-    assert_eq!(max_concurrent_jobs(&cfg), 64);
-
-    cfg.max_concurrent_jobs = Some(128);
-    assert_eq!(max_concurrent_jobs(&cfg), 64);
-
-    cfg.max_concurrent_jobs = Some(usize::MAX);
-    assert_eq!(max_concurrent_jobs(&cfg), 64);
-}
-
-#[test]
-fn phase_e2_polling_dispatch_and_job_execution_concurrency_defaults_are_independent() {
-    assert_eq!(POLLING_DISPATCH_MAX_IN_FLIGHT, 2);
-    assert_eq!(DEFAULT_MAX_CONCURRENT_JOBS, 4);
-    assert_ne!(POLLING_DISPATCH_MAX_IN_FLIGHT, DEFAULT_MAX_CONCURRENT_JOBS);
-}
+#[path = "main_tests/agent_config.rs"]
+mod agent_config;
+#[path = "main_tests/config_reload.rs"]
+mod config_reload;
+#[path = "main_tests/http_recovery.rs"]
+mod http_recovery;
 
 // ---------------------------------------------------------------------------
 // Stage 2G: platform-aware shell test helpers.
